🎉 init(repo): start my first VS Code extension
It is extremely frustrating when I'm coding and mistakenly touching the
touchpad, triggering a mouse-click event, resulting in a flight of the
cursor. I found an extension 'vscode-disable-mouse' and tried it, but it
has no effect with Vim extension on insert-mode. For the sake of my fun
and productivity, I'd like a try developing an extension of my own.

BTW I don't think this guy has tested it with Vim extension.

See: https://github.com/shio-yaamaa/vscode-disable-mouse